move stages to its own file
