+++ types/node_definition.go
@@ -76,140 +76,6 @@ type NodeDefinition struct {
 	HealthCheck *HealthcheckConfig `yaml:"healthcheck,omitempty"`
 }
 
-// Stages represents a configuration of a given node deployment stage.
-type Stages struct {
-	Create      *StageCreate      `yaml:"create"`
-	CreateLinks *StageCreateLinks `yaml:"create-links"`
-	Configure   *StageConfigure   `yaml:"configure"`
-	Healthy     *StageHealthy     `yaml:"healthy"`
-	Exit        *StageExit        `yaml:"exit"`
-}
-
-func NewStages() *Stages {
-	return &Stages{
-		Create: &StageCreate{
-			StageBase: StageBase{},
-		},
-		CreateLinks: &StageCreateLinks{
-			StageBase: StageBase{},
-		},
-		Configure: &StageConfigure{
-			StageBase: StageBase{},
-		},
-		Healthy: &StageHealthy{
-			StageBase: StageBase{},
-		},
-		Exit: &StageExit{
-			StageBase: StageBase{},
-		},
-	}
-}
-
-// GetWaitFor returns lists of nodes that need to be waited for in a map
-// that is indexed by the state for which this dependency is to be evaluated.
-func (s *Stages) GetWaitFor() map[WaitForPhase]WaitForList {
-	result := map[WaitForPhase]WaitForList{}
-
-	result[WaitForConfigure] = s.Configure.WaitFor
-	result[WaitForCreate] = s.Create.WaitFor
-	result[WaitForCreateLinks] = s.CreateLinks.WaitFor
-	result[WaitForHealthy] = s.Healthy.WaitFor
-	result[WaitForExit] = s.Exit.WaitFor
-
-	return result
-}
-
-// Merge merges the Stages configuration of other into s.
-func (s *Stages) Merge(other *Stages) error {
-	var err error
-	if other.Configure != nil {
-		err = s.Configure.Merge(&other.Configure.StageBase)
-		if err != nil {
-			return err
-		}
-	}
-	if other.Create != nil {
-		err = s.Create.Merge(&other.Create.StageBase)
-		if err != nil {
-			return err
-		}
-	}
-	if other.CreateLinks != nil {
-		err = s.CreateLinks.Merge(&other.CreateLinks.StageBase)
-		if err != nil {
-			return err
-		}
-	}
-	if other.Healthy != nil {
-		err = s.Healthy.Merge(&other.Healthy.StageBase)
-		if err != nil {
-			return err
-		}
-	}
-	if other.Exit != nil {
-		err = s.Exit.Merge(&other.Exit.StageBase)
-		if err != nil {
-			return err
-		}
-	}
-	return err
-}
-
-// StageCreate represents a creation stage of a given node.
-type StageCreate struct {
-	StageBase `yaml:",inline"`
-}
-
-// StageCreateLinks represents a stage of a given node when links are getting added to it.
-type StageCreateLinks struct {
-	StageBase `yaml:",inline"`
-}
-
-// StageConfigure represents a stage of a given node when it enters configuration workflow.
-type StageConfigure struct {
-	StageBase `yaml:",inline"`
-}
-
-// StageHealthy represents a stage of a given node when it reaches healthy status.
-type StageHealthy struct {
-	StageBase `yaml:",inline"`
-}
-
-// StageExit represents a stage of a given node when the node reaches exit state.
-type StageExit struct {
-	StageBase `yaml:",inline"`
-}
-
-// StageBase represents a configuration of a given stage.
-type StageBase struct {
-	WaitFor WaitForList `yaml:"wait-for,omitempty"`
-}
-
-type WaitForList []*WaitFor
-
-func (w WaitForList) contains(newWf *WaitFor) bool {
-	for _, entry := range w {
-		if entry.Equals(newWf) {
-			return true
-		}
-	}
-	return false
-}
-
-func (s *StageBase) Merge(sc *StageBase) error {
-	if sc == nil {
-		return nil
-	}
-	for _, wf := range sc.WaitFor {
-		// prevent adding the same dependency twice
-		if s.WaitFor.contains(wf) {
-			continue
-		}
-		s.WaitFor = append(s.WaitFor, wf)
-	}
-	return nil
-}
-
 // Interface compliance.
 var _ yaml.Unmarshaler = &NodeDefinition{}
